+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Changelog
 - Fixed hardware wallets confirmation crashing issue when device is in "Busy" state ([PR 2522](https://github.com/input-output-hk/daedalus/pull/2522))
 - Fixed hardware wallets spending password issue ([PR 2519](https://github.com/input-output-hk/daedalus/pull/2519))
 
+### Chores
+
+- Updated `@cardano-foundation/ledgerjs-hw-app-cardano` package to version `3.0.0` ([PR 2532](https://github.com/input-output-hk/daedalus/pull/2532))
+
 ## 4.0.5
 
 ### Features